requirements.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.7
backcall==0.2.0
black==20.8b1
click==7.1.2
colorama==0.4.3
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.6
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
django-import-export==2.3.0
et-xmlfile==1.0.1
heroku==0.1.4
gunicorn==20.0.4
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
idna==2.10
jdcal==1.4.1
mypy-extensions==0.4.3
MarkupPy==1.14
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
psycopg2==2.8.6
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2020.1
odfpy==1.4.1
openpyxl==3.0.5
regex==2020.7.14
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
